requirements.txt
Flask==2.2.3
tensorflow==2.12.0
keras==2.12.0
pillow==9.0.1
numpy==1.22.3
gunicorn==20.1.0
werkzeug==2.2.2
Flask-Cors==4.0.0